requirements.txt
mysql-connector-python
matplotlib
pandas
hashlib
